chore: Refactor Application Runtime.
